Suggesting improvements to the ansible test action.
At a minimum the ansible test job will need to include ansible-lint in the building steps.  

I also included other improvements I have in the state of the component in my instance. A little more descriptive info on what commands are running in the background and component variable descriptors, and I added ">/dev/null" to silence the output for tidiness.